requirements.txt
torch
torchvision
tqdm
Pillow
tk
